implement outside popup close function
